+++ webapp/src/App.tsx
@@ -5,9 +5,8 @@ import { FluentProvider, Subtitle1, makeStyles, shorthands, tokens } from '@flue
 
 import * as React from 'react';
 import { useEffect } from 'react';
-import { UserSettingsMenu } from './components/header/UserSettingsMenu';
-import { PluginGallery } from './components/open-api-plugins/PluginGallery';
-import { BackendProbe, ChatView, Error, Loading, Login } from './components/views';
+import Chat from './components/chat/Chat';
+import { Loading, Login } from './components/views';
 import { AuthHelper } from './libs/auth/AuthHelper';
 import { useChat, useFile } from './libs/hooks';
 import { AlertType } from './libs/models/AlertType';
@@ -47,7 +46,7 @@ export const useClasses = makeStyles({
     },
 });
 
-enum AppState {
+export enum AppState {
     ProbeForBackend,
     SettingUserInfo,
     ErrorLoadingChats,
@@ -154,57 +153,6 @@ const App = () => {
                 content
             )}
         </FluentProvider>
-    );
-};
-
-const Chat = ({
-    classes,
-    appState,
-    setAppState,
-}: {
-    classes: ReturnType<typeof useClasses>;
-    appState: AppState;
-    setAppState: (state: AppState) => void;
-}) => {
-    const onBackendFound = React.useCallback(() => {
-        setAppState(
-            AuthHelper.isAuthAAD()
-                ? // if AAD is enabled, we need to set the active account before loading chats
-                  AppState.SettingUserInfo
-                : // otherwise, we can load chats immediately
-                  AppState.LoadingChats,
-        );
-    }, [setAppState]);
-    return (
-        <div className={classes.container}>
-            <div className={classes.header}>
-                <Subtitle1 as="h1">Chat Copilot</Subtitle1>
-                {appState > AppState.SettingUserInfo && (
-                    <div className={classes.cornerItems}>
-                        <div className={classes.cornerItems}>
-                            <PluginGallery />
-                            <UserSettingsMenu
-                                setLoadingState={() => {
-                                    setAppState(AppState.SigningOut);
-                                }}
-                            />
-                        </div>
-                    </div>
-                )}
-            </div>
-            {appState === AppState.ProbeForBackend && <BackendProbe onBackendFound={onBackendFound} />}
-            {appState === AppState.SettingUserInfo && (
-                <Loading text={'Hang tight while we fetch your information...'} />
-            )}
-            {appState === AppState.ErrorLoadingUserInfo && (
-                <Error text={'Unable to load user info. Please try signing out and signing back in.'} />
-            )}
-            {appState === AppState.ErrorLoadingChats && (
-                <Error text={'Unable to load chats. Please try refreshing the page.'} />
-            )}
-            {appState === AppState.LoadingChats && <Loading text="Loading chats..." />}
-            {appState === AppState.Chat && <ChatView />}
-        </div>
     );
 };
